# Faster matching of events