--- a/tests/test_util.py
+++ b/tests/test_util.py
@@ -201,17 +201,6 @@ def test_outer_distance_mod_n():
     assert np.allclose(actual, expected)
 
 
-def test_outer_distance():
-    ref = [1., 2., 3.]
-    est = [1.1, 6., 1.9, 5., 10.]
-    expected = np.array([
-        [0.1, 5., 0.9, 4., 9.],
-        [0.9, 4., 0.1, 3., 8.],
-        [1.9, 3., 1.1, 2., 7.]])
-    actual = mir_eval.util._outer_distance(ref, est)
-    assert np.allclose(actual, expected)
-
-
 def test_match_events():
     ref = [1., 2., 3.]
     est = [1.1, 6., 1.9, 5., 10.]