Ths PR addresses #257 by doing a sorted bisection search for finding matched events, rather than building an all-pairs distance matrix.  The latter is still supported for fancy distance metrics, as is used in multipitch eval.

Quick benchmark:
```python

In [5]: est = np.random.randn(1000)

In [6]: ref = np.random.randn(10000) * 100

In [7]: %timeit mir_eval.util.match_events(ref, est, window=0.5)
57.4 ms ± 381 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [8]: %timeit mir_eval.util.match_events(ref, est, window=0.5, distance=mir_eval.util._outer_distance)
183 ms ± 1.52 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [9]: est = np.random.randn(10000)

In [10]: %timeit mir_eval.util.match_events(ref, est, window=0.5, distance=mir_eval.util._outer_distance)
1.88 s ± 46.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [11]: %timeit mir_eval.util.match_events(ref, est, window=0.5)
536 ms ± 5.13 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```