@lostanlen 2017-08-01
Looks good to me. Thank you for putting the time into this.