@bmcfee 2017-08-01
I put one more minor optimization in there, removing the inner-loop from `fast_hit_windows`.  I think this is now as fast as can be, and ready for CR.