@craffel 2017-08-01
Looks good, as long as all the regression tests pass :)

Can we remove `_outer_distance`?  Or just deprecate it?