@bmcfee 2017-08-01
> Can we remove _outer_distance? Or just deprecate it?

It's a private(ish) function that isn't referenced outside of util, so I'm cool with removing it.  Do you want me to put that in this PR as well?

I think it might be worth putting some thought into whether a similar trick could be done with `_outer_distance_modn` so we can get rid of all the quadratic-time calculations.  I don't immediately see an easy way to do it, but I also haven't thought about it too carefully.  Either way, that can be handled as a separate issue.