@craffel 2017-08-01
Thanks for this!